refactor: reuse meeting participation list component in treatment
