+++ app/components/meeting-form.hbs
@@ -124,6 +124,7 @@
             @aanwezigenBijStart={{this.aanwezigenBijStart}}
             @onSave={{this.saveParticipationList}}
             @bestuursorgaan={{this.bestuursorgaan}}
+            @modalTitle={{t "participationList.openModalButton"}}
           />
 
           <AuHeading id="sectionThree" @level="2" @skin="3" class="au-c-onboarding-wrapper">
