+++ app/components/participation-list.hbs
@@ -35,7 +35,7 @@
         {{on 'click' this.togglePopup}}
         @disabled={{this.fetchMandatees.isRunning}}
       >
-        {{t "participationList.openModalButton"}}
+        {{@modalTitle}}
       </AuButton>
 
       <ParticipationList::Modal
