--- app/components/treatment/participation.hbs
@@ -1,36 +1,0 @@
-<div class="au-c-meeting-chrome-card">
-  <ParticipationList::Modal
-    @show={{this.showParticipationModal}}
-    @aanwezigenBijStart={{this.aanwezigen}}
-    @possibleParticipants={{@possibleParticipants}}
-    @onSave={{this.saveParticipants}}
-    @togglePopup={{this.toggleModal}}
-    @voorzitter={{this.behandeling.voorzitter}}
-    @secretaris={{this.behandeling.secretaris}}
-    @bestuursorgaan={{@bestuursorgaan}}
-  />
-  {{#if this.loadData.isRunning}}
-    <p>{{t "participationList.loadingTitle"}}</p>
-    <p class="loader"><span class="u-visually-hidden">{{t "participationList.loadingLoader"}}</span></p>
-
-  {{else}}
-  <ul class="au-c-list-divider">
-    <li class="au-c-list-divider__item">
-      {{t "participationList.voorzitterLabel"}} <strong>{{this.voorzitter.isBestuurlijkeAliasVan.fullName}}</strong>
-    </li>
-
-    <li class="au-c-list-divider__item">
-      {{t "participationList.secretarisLabel"}} <strong>{{this.secretaris.isBestuurlijkeAliasVan.fullName}}</strong>
-    </li>
-    <li class="au-c-list-divider__item">
-      {{t "participation.presentLabel"}}
-      {{#each this.sortedAanwezigen as |mandataris|}}
-          <strong>{{mandataris.isBestuurlijkeAliasVan.fullName}}</strong>,
-      {{/each}}
-    </li>
-  </ul>
-  {{/if}}
-  <AuButton {{on "click" this.toggleModal}} @skin="secondary" @width="block" @icon="pencil" @iconAlignment="left">
-    {{t "behandelingVanAgendapunten.participationListButton"}}
-  </AuButton>
-</div>
